refactor: validator 의존성 수정
